requirements.txt
fastmcp>=2.12.0
uvicorn>=0.35.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
firecrawl-py>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
